requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
boto3==1.18.47
botocore==1.21.47
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
cloudinary==1.26.0
coverage==6.3.2
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==3.2.7
django-active-link==0.1.8
django-allauth==0.50.0
django-autoslug==1.9.8
django-bootstrap4==3.0.1
django-ckeditor==6.1.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
django-countries==7.2.1
django-credit-cards==0.4.1
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.4
django-decouple==2.0
django-email-verification==0.2.2
django-filter==21.1
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-oauth-toolkit==1.7.1
django-phone-field==1.8.1
django-phonenumber-field==6.1.0
django-social-share==2.2.1
django-storages==1.11.1
django-template-maths==0.2.0
django-tinymce==3.5.0
djangorestframework==3.13.1
google-api-core==2.0.1
google-auth==2.1.0
google-cloud-core==2.0.0
google-cloud-storage==1.42.2
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.11.3
install==1.3.4
jmespath==0.10.0
jwcrypto==1.0
Markdown==3.3.6
oauthlib==3.2.0
Pillow==8.3.2
protobuf==3.18.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.14.0
zipp==3.8.0
